+++ app/tests/test_dashboard.py
@@ -1,13 +1,7 @@
 import pytest
 
 from app import create_app
-
-@pytest.fixture
-def client():
-    app = create_app()
-    app.config['TESTING'] = True
-    client = app.test_client()
-    yield client
+from .utils import client
 
 def test_root(client):
     response = client.get('/dashboard')
