+++ app/tests/utils.py
@@ -1,0 +1,10 @@
+import pytest
+
+from app import create_app
+
+@pytest.fixture
+def client():
+    app = create_app()
+    app.config['TESTING'] = True
+    with app.test_client() as client:
+        yield client
